% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ordinary_krige.R, R/Universal_krige_covar_sqrt_dist.R
\name{tool_exec}
\alias{tool_exec}
\title{Developing a script that enables a tool to run the ordinary kriging}
\usage{
tool_exec(in_params, out_params)

tool_exec(in_params, out_params)
}
\arguments{
\item{in_params}{- input parameters}

\item{out_params}{- output parameters}

\item{input_feature}{-  Input point feature containing fields of the dependant variable and all explanatory variables.}

\item{predict_location}{-  Input point feature representing locations where you would like to predict the probable values for the presence of dependant variable. These point feature must have certain explanatory variables stored as fields.}

\item{dep_variable}{- Field from the input feature containing the sampled attributes. A particular value gives the strength of the field element at that point.}

\item{log_var}{- Taking logarithmic values for the dependent variable.}

\item{partial_sill}{- (Partial) Sill of the variogram model component.}

\item{modl}{- Model type, e.g. "Exp", "Sph", "Gau", "Mat".}

\item{rang}{- Range of the variogram model component.}

\item{nugt}{- Nugget component of the variogram. (this basically adds a nugget compontent to the model)}

\item{input_feature}{-  Input point feature containing fields of the dependant variable and all explanatory variables.}

\item{predict_location}{-  Input point feature representing locations where you would like to predict the probable values for the presence of dependant variable. These point feature must have certain explanatory variables stored as fields.}

\item{dep_variable}{- Field from the input feature containing the sampled attributes. A particular value gives the strength of the field element at that point.}

\item{log_var}{- Taking logarithmic values for the dependent variable.}

\item{covariate_var}{- Field from the input feature containing independent or explanatory variables that would act as a variable to covariate for the interpolation.}

\item{vgm_mod}{- Input for the vgm expression to fit the variogram (default : vgm(1,"Exp",300,1))}
}
\value{
output_feature1 - Output krige shapefile that contains the predictions of the values of unsampled locations from the Prediction_location dataset.

output_feature2 - Output variance provides how far the values are deviated from the other and the mean and exports the output as a pdf file.

output_feature1 -Output krige shapefile that contains the predictions of the values of unsampled locations from the Prediction_location dataset.

output_feature2 - Output variance provides how far the values are deviated from the other and the mean and exports the output as a pdf file.
}
\description{
This script function builds an interpolated result without taking into account of the covariates
and provide a kriged plot upon being run with the corresponding ArcGIS script tool.

This script function builds an interpolated result with the consideration of the covariate, particularly sqrt(dist)
and provide a kriged plot upon being run with the corresponding ArcGIS script tool.
}
\details{
\code{tool_exec}

This script function interpolates the values that are modeled by a Gaussian process governed by prior covariance,
as opposed to a piecewise–polynomial spline chosen to optimize smoothness of the fitted values.
it is particularly associated with linear unbiased estimation, as the estimations are weighed linear combination of available data.
It does not take into account of the covariates
i.e. it assumes constant unknown mean and helps in minimizing error variance.

This script function interpolates the values that are modeled by a Gaussian process governed by prior covariance,
as opposed to a piecewise–polynomial spline chosen to optimize smoothness of the fitted values.
it is similar to ordinary kriging, yet it includes a polynomial trend model,
assumes covariates for interpolation (kriging under non-stationary conditions, in the presence of drift).
}
\author{
Shankarlingam Sundaresan

Shankarlingam Sundaresan
}

