% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComKrige.R
\name{tool_exec}
\alias{tool_exec}
\title{Developing a script that enables a tool to run the ordinary and universal kriging, in case with square root of distance as the covariate}
\usage{
tool_exec(in_params, out_params)
}
\arguments{
\item{in_params}{- input parameters}

\item{out_params}{- output parameters}
}
\value{
output_feature1 - \code{\link[gstat]{krige}} Output krige shapefile that contains the predictions of the values of unsampled locations from the Prediction_location dataset.

output_feature2 - \code{\link[gstat]{variogram}} Output variance provides how far the values are deviated from the other and the mean and exports the output as a pdf file.
}
\description{
This script function builds an interpolated result with the consideration of the covariate in case if the covariate variable is given, particularly sqrt(dist)
and provide a kriged plot upon being run with the corresponding ArcGIS script tool.
}
\details{
This script function interpolates the values that are modeled by a Gaussian process governed by prior covariance, as opposed to a piecewise polynomial spline chosen to optimize smoothness of the fitted values.it is similar to ordinary kriging, yet it includes a polynomial trend model,
assumes covariates for interpolation (kriging under non-stationary conditions, in the presence of drift).
}
\note{
The in_params parameter includes a list of 6 arguments.

\strong{input_feature} -  Input point feature containing fields of the dependant variable and all explanatory variables.

\strong{predict_location} -  Input point feature representing locations where you would like to predict the probable values for the presence of dependant variable. These point feature must have certain explanatory variables stored as fields.

\strong{dep_variable} - Field from the input feature containing the sampled attributes. A particular value gives the strength of the field element at that point.

\strong{log_var} - Taking logarithmic values for the dependent variable.

\strong{covariate_var} - Field from the input feature containing independent or explanatory variables that would act as a variable to covariate for the interpolation.

\strong{vgm_mod} - \code{\link[gstat]{fit.variogram}}Input for the vgm expression to fit the variogram (\emph{default : vgm(1,"Exp",300,1) , which is meant for the \bold{universal kriging}})

The out_params parameter includes the return values of \emph{output_feature1} and \emph{output_feature2}
}
\author{
Edzer Pebesma, Shankarlingam Sundaresan
}
\seealso{
\code{\link{gstat}} , \code{\link{arcgisbinding}}
}

